refactor: extract Tibero routine trigger grant loader
